refactor: P2-4 split Extensions.swift into modular files
